[NCL-7650]: Compute only count of artifacts needed, not fetching from DB entirely

+++ dto/src/main/java/org/jboss/pnc/dto/response/statistics/ProductMilestoneStatistics.java
@@ -51,12 +51,12 @@ public class ProductMilestoneStatistics {
     /**
      * Proportion of quality of Delivered Artifacts.
      */
-    EnumMap<ArtifactQuality, Integer> artifactQuality;
+    EnumMap<ArtifactQuality, Long> artifactQuality;
 
     /**
      * Proportion of repository type of Delivered Artifacts.
      */
-    EnumMap<RepositoryType, Integer> repositoryType;
+    EnumMap<RepositoryType, Long> repositoryType;
 
     @JsonPOJOBuilder(withPrefix = "")
     @JsonIgnoreProperties(ignoreUnknown = true)
